refactor : setState에서 모든 state를 반영하도록 변경 / 파일별 관심사 분리
