+++ src/todoInput.jsx
@@ -1,35 +1,16 @@
-import React, { useState } from 'react';
+import React from 'react';
 
-export default function TodoInput({ appendTodoList }) {
-  const [state, setState] = useState({
-    input: '',
-  });
-
-  const { input } = state;
-
-  function updateInput(value) {
-    setState({
-      input: value,
-    });
-  }
-
-  function updateTodoList(value) {
-    appendTodoList(value);
-    setState({
-      input: '',
-    });
-  }
-
+export default function TodoInput({ todoInput, updateInput, updateTodoList }) {
   return (
     <div>
       <input
         onChange={(e) => updateInput(e.target.value)}
         placeholder="할 일을 입력해주세요."
-        value={input}
+        value={todoInput}
       />
       <button
         type="button"
-        onClick={() => updateTodoList(input)}
+        onClick={() => updateTodoList(todoInput)}
       >
         추가
       </button>
